@@ -3,18 +3,17 @@ import 'package:flutter/material.dart';
 class AuthBackground extends StatelessWidget {
   final Widget child;
 
-  const AuthBackground({Key key, @required this.child}) : super(key: key);
+  const AuthBackground({
+    Key? key,
+    required this.child,
+  }) : super(key: key);
   @override
   Widget build(BuildContext context) {
     return Container(
       width: double.infinity,
       height: double.infinity,
       child: Stack(
-        children: [
-          _PurpleBox(),
-          _HeaderIcon(),
-          this.child
-        ],
+        children: [_PurpleBox(), _HeaderIcon(), this.child],
       ),
     );
   }
