@@ -4,8 +4,8 @@ class CardContainer extends StatelessWidget {
   final Widget child;
 
   const CardContainer({
-    Key key, 
-    @required this.child
+    Key? key, 
+    required this.child
   }) : super(key: key);
 
   @override
