@@ -15,8 +15,8 @@ class LoginFormProvider extends ChangeNotifier {
   }
 
   bool isValidForm() {
-    print(formKey.currentState.validate());
+    print(formKey.currentState?.validate());
     print('$email - $password');
-    return formKey.currentState.validate() ?? false;
+    return formKey.currentState?.validate() ?? false;
   }
 }
