@@ -1,18 +1,33 @@
 import 'package:flutter/material.dart';
+import 'package:provider/provider.dart';
+
+import 'package:productos_app/screens/screens.dart';
+import 'package:productos_app/services/services.dart';
 import 'package:productos_app/widgets/widgets.dart';
 
 class HomeScreen extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
+    final productsService = Provider.of<ProductsService>(context);
+    if (productsService.isLoading) return LoadingScreen();
     return Scaffold(
       appBar: AppBar(
         title: Text('Productos'),
       ),
       body: ListView.builder(
-        itemCount: 10,
+        itemCount: productsService.products.length,
         itemBuilder: (BuildContext context, int index) => GestureDetector(
-          onTap: () => Navigator.pushNamed(context, 'product'),
-          child: ProductCard(),
+          onTap: () {
+            productsService.selectedProduct = productsService.products[index];
+            Navigator.pushNamed(
+              context,
+              'product',
+              arguments: productsService.products[index],
+            );
+          },
+          child: ProductCard(
+            product: productsService.products[index],
+          ),
         ),
       ),
       floatingActionButton: FloatingActionButton(
